@@ -14,16 +14,16 @@ ext.isSnapshot = false  //only true while in development
 ext.minimumSdkVersion = 19
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
         applicationId "com.job.darasalecturer"
         minSdkVersion 19
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode generateVersionCode() // 190020501
         versionName generateVersionName() // 2.5.1
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         //Enabling multidex support.
         multiDexEnabled true
@@ -62,6 +62,16 @@ android {
 
     packagingOptions {
         exclude 'META-INF/proguard/androidx-annotations.pro'
+        exclude 'META-INF/LICENSE'
+    }
+    compileOptions {
+        sourceCompatibility = '1.8'
+        targetCompatibility = '1.8'
+    }
+
+    dexOptions {
+        //may speed up incremental builds
+        preDexLibraries true
     }
 
 }
@@ -86,20 +96,21 @@ dependencies {
     def support_version = "28.0.0-alpha1"
     def lifecycle_version = "1.1.1"
     def room_version = "1.1.1"
+    def work_version = "2.2.0"
     def firebase_common = "16.1.0"
-    def butterknife_version = "8.8.1"
+    def butterknife_version = "10.0.0"
     def lottie_version = "2.7.0"
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "com.android.support:appcompat-v7:$support_version"
-    implementation "com.android.support:support-vector-drawable:$support_version"
-    implementation "com.android.support:support-v4:$support_version"
-    implementation "com.android.support:design:$support_version"
-    implementation "com.android.support:cardview-v7:$support_version"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    //implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     //If your minSdkVersion is lower than 21
-    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'androidx.multidex:multidex:2.0.1'
 
     //firebase
     implementation 'com.google.firebase:firebase-core:16.0.6'
@@ -119,7 +130,7 @@ dependencies {
     implementation 'com.firebaseui:firebase-ui-firestore:4.1.0'
 
     //rating bar
-    implementation 'me.zhanghai.android.materialratingbar:library:1.2.0'
+    implementation 'me.zhanghai.android.materialratingbar:library:1.3.2'
 
     //sweet alert dialogue
     implementation 'com.github.f0ris.sweetalert:library:1.5.1'
@@ -166,15 +177,17 @@ dependencies {
     implementation "com.github.skydoves:powermenu:2.0.5"
 
     //work manager
-    //implementation "android.arch.work:work-runtime:1.0.0-alpha10"
-    implementation("android.arch.work:work-runtime:1.0.0-alpha10") {
+    // (Java only)
+    implementation "androidx.work:work-runtime:$work_version"
+
+    /*implementation('androidx.work:work-runtime:2.0.1') {
         exclude group: 'com.google.guava', module: 'listenablefuture'
-    }
+    }*/
 
     //architecture components
     // ViewModel and LiveData
-    implementation "android.arch.lifecycle:extensions:$lifecycle_version"
-    annotationProcessor "android.arch.lifecycle:compiler:$lifecycle_version"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
+    annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.1.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
