@@ -1,10 +1,10 @@
 package com.job.darasalecturer.ui.newlesson;
 
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProviders;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.job.darasalecturer.R;
 import com.job.darasalecturer.adapter.AddClassPagerAdapter;
