@@ -1,8 +1,8 @@
 package com.job.darasalecturer.viewmodel;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MediatorLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MediatorLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.job.darasalecturer.model.CourseYear;
 import com.job.darasalecturer.model.LecTeach;
