@@ -1,7 +1,7 @@
 package com.job.darasalecturer.viewmodel;
 
-import android.arch.lifecycle.MediatorLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.MediatorLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.job.darasalecturer.model.StudentDetails;
 
