@@ -1,7 +1,7 @@
 package com.job.darasalecturer.viewmodel;
 
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.job.darasalecturer.model.LecTeach;
 
